requirements.txt
alembic==1.5.6
aniso8601==9.0.1
attrs==20.3.0
autopep8==1.5.5
cffi==1.14.5
click==7.1.2
cryptography==3.4.6
Flask==1.1.2
Flask-Migrate==2.7.0
flask-restplus==0.13.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
importlib-metadata==3.7.2
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
Mako==1.1.4
MarkupSafe==1.1.1
pycodestyle==2.6.0
pycparser==2.20
PyJWT==2.0.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
typing-extensions==3.7.4.3
Werkzeug==0.16.0
zipp==3.4.1
